refactor: move field collision detection to server
- For consistency with model/taxonomy name collision detection.
- To make it easier to run field collision detection when fields are
created outside of our field UI.
- To simplify upcoming work to allow 'title' as a field id.
