+++ includes/rest-api/graphql.php
@@ -75,6 +75,81 @@ const ACM_DEFAULT_GRAPHQL_ROOT_FIELDS = [
 ];
 
 /**
+ * Reserved field names in use by WPGraphQL for the WordPress Core “Post” type.
+ *
+ * We can't use these values as field names because an ACM field named “title”
+ * would conflict with the WP/WPGraphQL “title”, for example.
+ *
+ * We report conflicts instead of namespacing ACM fields under an `acmFields`
+ * group to GraphQL responses to improve developer ergonomics: every field is a
+ * top-level property of its model and can be used alongside built-in reserved
+ * properties without namespacing.
+ *
+ * Default reserved names are derived from the WPGraphQL response to this query:
+ *
+ * ```
+ * query GetTypeAndFields {
+ *     __type(name: "Post" ) {
+ *         fields {
+ *             name
+ *         }
+ *     }
+ * }
+ * ```
+ */
+const ACM_RESERVED_GRAPHQL_FIELD_IDS = [
+	'author',
+	'authorDatabaseId',
+	'authorId',
+	'categories',
+	'commentCount',
+	'commentStatus',
+	'comments',
+	'content',
+	'contentType',
+	'databaseId',
+	'date',
+	'dateGmt',
+	'desiredSlug',
+	'editingLockedBy',
+	'enclosure',
+	'enqueuedScripts',
+	'enqueuedStylesheets',
+	'excerpt',
+	'featuredImage',
+	'featuredImageDatabaseId',
+	'featuredImageId',
+	'guid',
+	'id',
+	'isContentNode',
+	'isPreview',
+	'isRestricted',
+	'isRevision',
+	'isSticky',
+	'isTermNode',
+	'lastEditedBy',
+	'link',
+	'modified',
+	'modifiedGmt',
+	'pingStatus',
+	'pinged',
+	'postFormats',
+	'preview',
+	'previewRevisionDatabaseId',
+	'previewRevisionId',
+	'revisionOf',
+	'revisions',
+	'slug',
+	'status',
+	'tags',
+	'template',
+	'terms',
+	'title',
+	'toPing',
+	'uri',
+];
+
+/**
  * Determines if the passed `$name` already exists in the root
  * WPGraphQL namespace.
  *
@@ -85,6 +160,18 @@ function root_type_exists( string $name = '' ): bool {
 	$root_fields = get_registered_root_fields();
 
 	return in_array( camelcase( $name ), $root_fields, true );
+}
+
+/**
+ * Determines if the passed field `$id` is in use in the given `$model`.
+ *
+ * @param string $id Field name to check for collisions with existing fields.
+ * @param string $model The case-sensitive WPGraphQL type.
+ *                      (Singular form, camel case, initial capital).
+ * @return bool True if the passed `$name` collides with a reserved field ID.
+ */
+function is_registered_field_id( string $id = '', string $model = '' ): bool {
+	return in_array( $id, get_registered_field_ids( $model ), true );
 }
 
 /**
@@ -128,4 +215,50 @@ function get_registered_root_fields(): array {
 	);
 
 	return is_array( $field_names ) ? $field_names : [];
+}
+
+/**
+ * Gets field IDs already in use for the given `$model` by making a GraphQL
+ * request using introspection to find current registered field names.
+ *
+ * When WPGraphQL is available, this allows us to check for conflicts against
+ * additional fields registered at runtime by WPGraphQL or other plugins.
+ * Falls back to `ACM_RESERVED_GRAPHQL_FIELD_IDS` when WPGraphQL is not active.
+ *
+ * Note that the response includes names of existing ACM fields, which are not
+ * currently filterable with WPGraphQL introspection, and not just fields
+ * registered by WPGraphQL itself and other plugins.
+ *
+ * @param string $model The case-sensitive WPGraphQL type.
+ *                      (Singular form, camel case, initial capital).
+ * @return array Reserved field ids for the given `$model`.
+ */
+function get_registered_field_ids( string $model = '' ): array {
+	if ( ! function_exists( 'graphql' ) || empty( $model ) ) {
+		// We can't use introspection so fall back to default field IDs.
+		return ACM_RESERVED_GRAPHQL_FIELD_IDS;
+	}
+
+	$registered_fields = graphql(
+		[
+			'query'     => '
+				query GetTypeAndFields($model: String!) {
+					__type(name: $model) {
+					fields {
+						name
+						}
+					}
+				}
+		  ',
+			'variables' => [
+				'model' => $model,
+			],
+		]
+	);
+
+	// The names of all fields.
+	return array_column(
+		$registered_fields['data']['__type']['fields'] ?? [],
+		'name'
+	);
 }
