+++ includes/rest-api/routes/content-model-field.php
@@ -11,6 +11,7 @@ namespace WPE\AtlasContentModeler\REST_API\ContentModelField;
 
 use WP_Error;
 use WP_REST_Request;
+use function WPE\AtlasContentModeler\ContentRegistration\camelcase;
 use function WPE\AtlasContentModeler\ContentRegistration\get_registered_content_types;
 use function WPE\AtlasContentModeler\ContentRegistration\update_registered_content_types;
 use function WPE\AtlasContentModeler\REST_API\Fields\shape_field_args;
@@ -18,6 +19,7 @@ use function WPE\AtlasContentModeler\REST_API\Fields\content_model_field_exists;
 use function WPE\AtlasContentModeler\REST_API\Fields\content_model_multi_option_exists;
 use function WPE\AtlasContentModeler\REST_API\Fields\content_model_multi_option_slug_exists;
 use function WPE\AtlasContentModeler\REST_API\Fields\content_model_reverse_slug_exists;
+use function WPE\AtlasContentModeler\REST_API\GraphQL\is_registered_field_id;
 
 add_action( 'rest_api_init', __NAMESPACE__ . '\register_rest_routes' );
 /**
@@ -59,9 +61,8 @@ function register_rest_routes(): void {
  * @return WP_Error|\WP_HTTP_Response|\WP_REST_Response
  */
 function dispatch_update_content_model_field( WP_REST_Request $request ) {
-	$params               = $request->get_params();
-	$content_types        = get_registered_content_types();
-	$reserved_field_slugs = include ATLAS_CONTENT_MODELER_INCLUDES_DIR . 'settings/reserved-field-slugs.php';
+	$params        = $request->get_params();
+	$content_types = get_registered_content_types();
 
 	if ( ! isset( $params['model'] ) || empty( $content_types[ $params['model'] ] ) ) {
 		return new WP_Error(
@@ -71,11 +72,13 @@ function dispatch_update_content_model_field( WP_REST_Request $request ) {
 		);
 	}
 
+	$graphql_type = ucfirst( camelcase( $content_types[ $params['model'] ]['singular'] ) );
+
 	// Prevents use of reserved field slugs during new field creation.
 	if (
 		$request->get_method() === 'POST' &&
 		isset( $params['slug'] ) &&
-		in_array( $params['slug'], $reserved_field_slugs, true )
+		is_registered_field_id( $params['slug'], $graphql_type, true )
 	) {
 		return new WP_Error(
 			'acm_reserved_field_slug',
