+++ includes/settings/js/src/hooks/index.js
@@ -1,2 +1,1 @@
 export { useInputGenerator } from "./useInputGenerator";
-export { useReservedSlugs } from "./useReservedSlugs";
